requirements.txt
anthropic>=0.18.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
pgvector>=0.2.0
python-dotenv>=1.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
schedule>=1.2.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
certifi>=2024.0.0
